Created separate UI comps, styling changes
